Refactor MediaMonkey test fixtures and cleanup tests
Because:
- MediaMonkey fixtures were duplicated across test files
- Tests contained excessive verbose documentation
- Test setup code was scattered and hard to maintain

Moves comprehensive MediaMonkey mocking infrastructure from test_mediamonkey_player.py to conftest.py for reusability. Removes verbose docstrings and comments to focus tests on essential assertions. Consolidates SQLQueryProcessor, SDBMock, and factory fixtures into shared conftest.py. Affects tests/conftest.py and tests/mediaplayer/test_mediamonkey_player.py with no functional changes to test coverage.
